backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic==2.11.3
pandas==2.2.3
python-dotenv==1.1.0
tenacity==9.1.2
beautifulsoup4==4.13.3
pytest==8.3.5

